AddedLocalStrategy

+++ controllers/controller.js
@@ -10,7 +10,8 @@ const index = async (req, res) => {
                    EXTRACT(DAY FROM posts.created_at) AS day
             FROM posts 
             JOIN users ON posts.user_id = users.id
-             WHERE users.flag = true
+             WHERE users.flag = true 
+             ORDER BY posts.id DESC
           `);
         const posts = result.rows;
         res.render("index.ejs", { posts: posts, user: req.user });
@@ -40,21 +41,22 @@ const post = async (req, res) => {
 }
 
 const create = (req, res) => {
-        const post = { id: 0, author: "", title: "", subtitle: "", content: "" };
-        res.render("create.ejs", { post: post, user: req.user });
+    const post = { id: 0, author: "", title: "", subtitle: "", content: "" };
+    res.render("create.ejs", { post: post, user: req.user });
 }
 
 const del = async (req, res) => {
-        const Id = req.params.post_id;
-        // Find the index of the post with the given ID
-        try {
-            const result = await db.query("SELECT * FROM posts WHERE id = $1", [Id]);
-            const found = result.rows;
-            if (found.length === 0) {
-                // No post found
-                res.status(404).send("No post with this id found.");
-            } else {
-                // If a post with the given ID is found, remove it
+    const Id = req.params.post_id;
+    // Find the index of the post with the given ID
+    try {
+        const result = await db.query("SELECT * FROM posts WHERE id = $1", [Id]);
+        const found = result.rows;
+        if (found.length === 0) {
+            // No post found
+            res.status(404).send("No post with this id found.");
+        } else {
+            // If a post with the given ID is found, remove it
+            if (req.user.id === found[0].user_id) {
                 try {
                     db.query("DELETE FROM posts WHERE id = $1", [Id]);
                     res.redirect("/");
@@ -62,23 +64,27 @@ const del = async (req, res) => {
                     console.log(err);
                     res.status(501).send("Error deleting post");
                 }
+            } else {
+                res.status(500).send("UnAuthorized User");
             }
-        } catch (err) {
-            console.log(err);
-            res.status(501).send("Error finding the post.");
         }
+    } catch (err) {
+        console.log(err);
+        res.status(501).send("Error finding the post.");
+    }
 }
 
 const edit = async (req, res) => {
-        const Id = req.params.post_id;
-        try {
-            const result = await db.query("SELECT * FROM posts WHERE id = $1", [Id]);
-            const found = result.rows;
-            if (found.length === 0) {
-                // No post found
-                res.status(404).send("Post not found.");
-            } else {
-                // If a post with the given ID is found, get the post details from database
+    const Id = req.params.post_id;
+    try {
+        const result = await db.query("SELECT * FROM posts WHERE id = $1", [Id]);
+        const found = result.rows;
+        if (found.length === 0) {
+            // No post found
+            res.status(404).send("Post not found.");
+        } else {
+            // If a post with the given ID is found, get the post details from database
+            if (req.user.id === found[0].user_id) {
                 const post = {
                     id: found[0].id,
                     title: found[0].title,
@@ -92,35 +98,38 @@ const edit = async (req, res) => {
                     imageurl: found[0].imageurl,
                 }
                 res.render("create.ejs", { post: post, user: req.user });
+            } else {
+                res.status(500).send("UnAuthorized User");
             }
-        } catch (err) {
-            console.log(err);
-            res.status(500).send("Error deleting the post.");
         }
+    } catch (err) {
+        console.log(err);
+        res.status(500).send("Error deleting the post.");
+    }
 }
 
 const create_post = async (req, res) => {
-        if (req.body.id == 0) {
-            try {
-                await db.query("INSERT INTO posts (title, subtitle, author, content, imageurl, user_id) VALUES ($1, $2, $3, $4, $5, $6)", [req.body.title, req.body.subtitle, req.user.username, req.body.content, req.body.url, req.user.id]);
-            } catch (err) {
-                if (err.code === "23505") {
-                    res.status(409).send("Post already exists with the same unique value.");
-                } else {
-                    console.log(err);
-                    res.status(500).send("Error inserting values into the database.");
-                }
-            }
-        } else {
-            try {
-                // Update the record if id is defined
-                await db.query("UPDATE posts SET title = $1, subtitle = $2, author = $3, content = $4, imageurl = $5, user_id = $6 WHERE id = $7", [req.body.title, req.body.subtitle, req.user.username, req.body.content, req.body.url, req.user.id, req.body.id]);
-            } catch (err) {
+    if (req.body.id == 0) {
+        try {
+            await db.query("INSERT INTO posts (title, subtitle, author, content, imageurl, user_id) VALUES ($1, $2, $3, $4, $5, $6)", [req.body.title, req.body.subtitle, req.user.username, req.body.content, req.body.url, req.user.id]);
+        } catch (err) {
+            if (err.code === "23505") {
+                res.status(409).send("Post already exists with the same unique value.");
+            } else {
                 console.log(err);
-                res.send("Error occurred whilst updating.");
+                res.status(500).send("Error inserting values into the database.");
             }
         }
-        res.redirect("/");
+    } else {
+        try {
+            // Update the record if id is defined
+            await db.query("UPDATE posts SET title = $1, subtitle = $2, author = $3, content = $4, imageurl = $5, user_id = $6 WHERE id = $7", [req.body.title, req.body.subtitle, req.user.username, req.body.content, req.body.url, req.user.id, req.body.id]);
+        } catch (err) {
+            console.log(err);
+            res.send("Error occurred whilst updating.");
+        }
+    }
+    res.redirect("/");
 }
 
 const contact = (req, res) => {
